rename chains.ts to xcmChains.ts
