+++ packages/config/src/ConfigBuilder/ConfigBuilder.test.ts
@@ -1,11 +1,11 @@
 /* eslint-disable sort-keys */
-import { Ecosystem } from '@moonbeam-network/xcm-types';
 import { describe, expect, it } from '@jest/globals';
+import { Ecosystem } from '@moonbeam-network/xcm-types';
 import { ConfigService } from '../ConfigService';
 import { dev } from '../assets';
-import { moonbaseAlpha, pendulumAlphanet } from '../chains';
 import { moonbaseAlphaConfig } from '../configs/moonbaseAlpha';
 import { pendulumAlphanetConfig } from '../configs/pendulumAlphanet';
+import { moonbaseAlpha, pendulumAlphanet } from '../xcmChains';
 import { ConfigBuilder } from './ConfigBuilder';
 
 describe('configBuilder', () => {
