+++ packages/config/src/configs/integritee.ts
@@ -4,9 +4,9 @@ import {
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
 import { teer } from '../assets';
-import { integritee, moonriver } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
+import { integritee, moonriver } from '../xcmChains';
 
 export const integriteeConfig = new ChainConfig({
   assets: [
