+++ packages/config/src/configs/kusama.ts
@@ -4,9 +4,9 @@ import {
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
 import { ksm } from '../assets';
-import { kusama, moonriver } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
+import { kusama, moonriver } from '../xcmChains';
 
 export const kusamaConfig = new ChainConfig({
   assets: [
