+++ packages/config/src/configs/picasso.ts
@@ -4,9 +4,9 @@ import {
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
 import { movr, pica } from '../assets';
-import { moonriver, picasso } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
+import { moonriver, picasso } from '../xcmChains';
 
 export const picassoConfig = new ChainConfig({
   assets: [
