+++ packages/config/src/configs/robonomics.ts
@@ -4,9 +4,9 @@ import {
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
 import { xrt } from '../assets';
-import { moonriver, robonomics } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
+import { moonriver, robonomics } from '../xcmChains';
 
 export const robonomicsConfig = new ChainConfig({
   assets: [
