+++ packages/config/src/configs/shiden.ts
@@ -4,9 +4,9 @@ import {
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
 import { movr, sdn } from '../assets';
-import { moonriver, shiden } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
+import { moonriver, shiden } from '../xcmChains';
 
 export const shidenConfig = new ChainConfig({
   assets: [
