+++ packages/config/src/configs/uniqueAlpha.ts
@@ -1,5 +1,5 @@
-import { uniqueAlpha } from '../chains';
 import { ChainConfig } from '../types/ChainConfig';
+import { uniqueAlpha } from '../xcmChains';
 
 // NOTE: Disabling because ws endpoint is not working
 export const uniqueAlphaConfig = new ChainConfig({
