+++ packages/config/src/index.ts
@@ -1,6 +1,6 @@
+export * from './assets';
 export * from './ConfigBuilder';
 export * from './ConfigService';
-export * from './assets';
-export * from './chains';
 export * from './types/AssetConfig';
 export * from './types/ChainConfig';
+export * from './xcmChains';
